requirements.txt
flask
azure-cognitiveservices-vision-face
pillow
